It still doesn't work unfortunately but I'm gonna mess around with the Pooper.
I added the Zero Hopper command which just sets the speed to zero once the up button on the D-Pad is released.
I also messed around with sensor stuff and got it to work, kinda? I need to go over the actual subsystem code however.
If someone could look it over and tell me what should be added to get the simulation to work well, please let me know.