Added controller support
I added button binding for the shooter subsystem and also I made a new command for shooting with the subsystem. By the way, should I get rid of the DriveShooterWithConstant command and replace it with just a simple setSpeed(); method? (which I think is in the Shooter.java subsystem). I would love some feedback.